tavus/requirements.txt
livekit-agents[openai,silero,turn-detector,deepgram,google,anthropic,cartesia,elevenlabs,rime,playai,groq,tavus,gladia,hedra,inworld,google,sarvam]
livekit-plugins-aws[realtime]
livekit-plugins-noise-cancellation
python-dotenv
requests>=2.32.0
annoy
pydantic
flask
pandas
websockets>=11.0.3
rich
mcp
librosa
moondream
fastapi
uvicorn